+++ src/main/java/net/geforcemods/securitycraft/blockentities/DisguisableBlockEntity.java
@@ -1,11 +1,13 @@
 package net.geforcemods.securitycraft.blockentities;
 
+import net.geforcemods.securitycraft.ClientHandler;
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.api.CustomizableBlockEntity;
 import net.geforcemods.securitycraft.api.Option;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.models.DisguisableDynamicBakedModel;
 import net.geforcemods.securitycraft.network.client.RefreshDisguisableModel;
+import net.minecraft.block.Blocks;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntityType;
 import net.minecraftforge.client.model.data.IModelData;
@@ -21,16 +23,40 @@ public class DisguisableBlockEntity extends CustomizableBlockEntity {
 	public void onModuleInserted(ItemStack stack, ModuleType module) {
 		super.onModuleInserted(stack, module);
 
-		if (!level.isClientSide && module == ModuleType.DISGUISE)
-			SecurityCraft.channel.send(PacketDistributor.ALL.noArg(), new RefreshDisguisableModel(worldPosition, true, stack));
+		if (module == ModuleType.DISGUISE) {
+			if (!level.isClientSide)
+				SecurityCraft.channel.send(PacketDistributor.ALL.noArg(), new RefreshDisguisableModel(worldPosition, true, stack));
+			else
+				ClientHandler.putDisguisedBeRenderer(this, stack);
+		}
 	}
 
 	@Override
 	public void onModuleRemoved(ItemStack stack, ModuleType module) {
 		super.onModuleRemoved(stack, module);
 
-		if (!level.isClientSide && module == ModuleType.DISGUISE)
-			SecurityCraft.channel.send(PacketDistributor.ALL.noArg(), new RefreshDisguisableModel(worldPosition, false, stack));
+		if (module == ModuleType.DISGUISE) {
+			if (!level.isClientSide)
+				SecurityCraft.channel.send(PacketDistributor.ALL.noArg(), new RefreshDisguisableModel(worldPosition, false, stack));
+			else
+				ClientHandler.DISGUISED_BLOCK_RENDER_DELEGATE.removeDelegateOf(this);
+		}
+	}
+
+	@Override
+	public void onLoad() {
+		super.onLoad();
+
+		if (level.isClientSide)
+			ClientHandler.putDisguisedBeRenderer(this, getModule(ModuleType.DISGUISE));
+	}
+
+	@Override
+	public void setRemoved() {
+		super.setRemoved();
+
+		if (level.isClientSide)
+			ClientHandler.DISGUISED_BLOCK_RENDER_DELEGATE.removeDelegateOf(this);
 	}
 
 	@Override
@@ -47,6 +73,6 @@ public class DisguisableBlockEntity extends CustomizableBlockEntity {
 
 	@Override
 	public IModelData getModelData() {
-		return new ModelDataMap.Builder().withInitial(DisguisableDynamicBakedModel.DISGUISED_BLOCK_RL, getBlockState().getBlock().getRegistryName()).build();
+		return new ModelDataMap.Builder().withInitial(DisguisableDynamicBakedModel.DISGUISED_STATE_RL, Blocks.AIR.defaultBlockState()).build();
 	}
 }