+++ src/main/java/net/geforcemods/securitycraft/inventory/ModuleItemContainer.java
@@ -13,6 +13,7 @@ public class ModuleItemContainer implements IInventory {
 	public final int size = 1;
 	private final ItemStack module;
 	public NonNullList<ItemStack> moduleInventory;
+	private DisguiseModuleMenu menu;
 
 	public ModuleItemContainer(ItemStack moduleItem) {
 		module = moduleItem;
@@ -85,6 +86,7 @@ public class ModuleItemContainer implements IInventory {
 	@Override
 	public ItemStack removeItemNoUpdate(int index) {
 		ItemStack stack = getItem(index);
+
 		setItem(index, ItemStack.EMPTY);
 		return stack;
 	}
@@ -112,6 +114,9 @@ public class ModuleItemContainer implements IInventory {
 		}
 
 		writeToNBT(module.getTag());
+
+		if (menu != null)
+			menu.slotsChanged(this);
 	}
 
 	@Override
@@ -141,5 +146,13 @@ public class ModuleItemContainer implements IInventory {
 		}
 
 		return true;
+	}
+
+	public void setMenu(DisguiseModuleMenu menu) {
+		this.menu = menu;
+	}
+
+	public ItemStack getModule() {
+		return module;
 	}
 }