+++ src/main/java/net/geforcemods/securitycraft/models/DisguisableDynamicBakedModel.java
@@ -9,6 +9,7 @@ import javax.annotation.Nonnull;
 import net.geforcemods.securitycraft.blocks.DisguisableBlock;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
+import net.minecraft.block.Blocks;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.model.BakedQuad;
 import net.minecraft.client.renderer.model.IBakedModel;
@@ -16,36 +17,32 @@ import net.minecraft.client.renderer.model.ItemOverrideList;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.Direction;
-import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.ILightReader;
 import net.minecraftforge.client.model.data.IDynamicBakedModel;
 import net.minecraftforge.client.model.data.IModelData;
 import net.minecraftforge.client.model.data.ModelProperty;
-import net.minecraftforge.registries.ForgeRegistries;
 
 public class DisguisableDynamicBakedModel implements IDynamicBakedModel {
-	public static final ModelProperty<ResourceLocation> DISGUISED_BLOCK_RL = new ModelProperty<>();
-	private final ResourceLocation defaultStateRl;
+	public static final ModelProperty<BlockState> DISGUISED_STATE_RL = new ModelProperty<>();
 	private final IBakedModel oldModel;
 
-	public DisguisableDynamicBakedModel(ResourceLocation defaultStateRl, IBakedModel oldModel) {
-		this.defaultStateRl = defaultStateRl;
+	public DisguisableDynamicBakedModel(IBakedModel oldModel) {
 		this.oldModel = oldModel;
 	}
 
 	@Override
 	public List<BakedQuad> getQuads(BlockState state, Direction side, Random rand, IModelData modelData) {
-		ResourceLocation rl = modelData.getData(DISGUISED_BLOCK_RL);
+		BlockState disguisedState = modelData.getData(DISGUISED_STATE_RL);
 
-		if (rl != null && rl != defaultStateRl) {
-			Block block = ForgeRegistries.BLOCKS.getValue(rl);
+		if (disguisedState != null) {
+			Block block = disguisedState.getBlock();
 
-			if (block != null) {
-				final IBakedModel model = Minecraft.getInstance().getBlockRenderer().getBlockModel(block.defaultBlockState());
+			if (block != Blocks.AIR) {
+				final IBakedModel model = Minecraft.getInstance().getBlockRenderer().getBlockModel(disguisedState);
 
 				if (model != null && model != this)
-					return model.getQuads(block.defaultBlockState(), side, rand, modelData);
+					return model.getQuads(disguisedState, side, rand, modelData);
 			}
 		}
 
@@ -54,13 +51,17 @@ public class DisguisableDynamicBakedModel implements IDynamicBakedModel {
 
 	@Override
 	public TextureAtlasSprite getParticleTexture(IModelData modelData) {
-		ResourceLocation rl = modelData.getData(DISGUISED_BLOCK_RL);
+		BlockState state = modelData.getData(DISGUISED_STATE_RL);
 
-		if (rl != null && rl != defaultStateRl) {
-			Block block = ForgeRegistries.BLOCKS.getValue(rl);
+		if (state != null) {
+			Block block = state.getBlock();
 
-			if (block != null && !(block instanceof DisguisableBlock))
-				return Minecraft.getInstance().getBlockRenderer().getBlockModel(block.defaultBlockState()).getParticleTexture(modelData);
+			if (block != Blocks.AIR) {
+				IBakedModel model = Minecraft.getInstance().getBlockRenderer().getBlockModel(state);
+
+				if (model != null && model != this)
+					return model.getParticleTexture(modelData);
+			}
 		}
 
 		return oldModel.getParticleTexture(modelData);
@@ -78,13 +79,13 @@ public class DisguisableDynamicBakedModel implements IDynamicBakedModel {
 				Optional<BlockState> disguisedState = ((DisguisableBlock) block).getDisguisedBlockState(world, pos);
 
 				if (disguisedState.isPresent()) {
-					tileData.setData(DISGUISED_BLOCK_RL, disguisedState.get().getBlock().getRegistryName());
+					tileData.setData(DISGUISED_STATE_RL, disguisedState.get());
 					return tileData;
 				}
 			}
 		}
 
-		tileData.setData(DISGUISED_BLOCK_RL, defaultStateRl);
+		tileData.setData(DISGUISED_STATE_RL, Blocks.AIR.defaultBlockState());
 		return tileData;
 	}