+++ src/main/java/net/geforcemods/securitycraft/screen/ProjectorScreen.java
@@ -1,6 +1,8 @@
 package net.geforcemods.securitycraft.screen;
 
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 
@@ -10,10 +12,13 @@ import net.geforcemods.securitycraft.inventory.ProjectorMenu;
 import net.geforcemods.securitycraft.network.server.SyncProjector;
 import net.geforcemods.securitycraft.network.server.SyncProjector.DataType;
 import net.geforcemods.securitycraft.screen.components.NamedSlider;
+import net.geforcemods.securitycraft.screen.components.StateSelector;
 import net.geforcemods.securitycraft.screen.components.StringHoverChecker;
 import net.geforcemods.securitycraft.screen.components.TogglePictureButton;
+import net.geforcemods.securitycraft.util.IHasExtraAreas;
 import net.geforcemods.securitycraft.util.Utils;
 import net.minecraft.client.gui.screen.inventory.ContainerScreen;
+import net.minecraft.client.renderer.Rectangle2d;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
@@ -21,10 +26,10 @@ import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 
 @OnlyIn(Dist.CLIENT)
-public class ProjectorScreen extends ContainerScreen<ProjectorMenu> {
+public class ProjectorScreen extends ContainerScreen<ProjectorMenu> implements IHasExtraAreas {
 	private static final ResourceLocation TEXTURE = new ResourceLocation("securitycraft:textures/gui/container/projector.png");
 	private static final String SLOT_TOOLTIP = Utils.localize("gui.securitycraft:projector.block").getColoredString();
-	private ProjectorBlockEntity tileEntity;
+	private ProjectorBlockEntity be;
 	private String blockName;
 	private StringHoverChecker[] hoverCheckers = new StringHoverChecker[5];
 	private StringHoverChecker slotHoverChecker;
@@ -32,55 +37,60 @@ public class ProjectorScreen extends ContainerScreen<ProjectorMenu> {
 	private NamedSlider projectionHeightSlider;
 	private NamedSlider projectionRangeSlider;
 	private NamedSlider projectionOffsetSlider;
+	private StateSelector stateSelector;
 	private int sliderWidth = 120;
 
-	public ProjectorScreen(ProjectorMenu container, PlayerInventory inv, ITextComponent name) {
-		super(container, inv, name);
-		this.tileEntity = container.te;
-		blockName = Utils.localize(tileEntity.getBlockState().getBlock().getDescriptionId()).getColoredString();
-		imageHeight = 225;
+	public ProjectorScreen(ProjectorMenu menu, PlayerInventory inv, ITextComponent text) {
+		super(menu, inv, text);
+		this.be = menu.be;
+		blockName = Utils.localize(be.getBlockState().getBlock().getDescriptionId()).getColoredString();
+		imageHeight = 235;
 	}
 
 	@Override
 	public void init() {
 		super.init();
+		leftPos += 90;
 
 		int id = 0;
 		int left = leftPos + ((imageWidth - sliderWidth) / 2);
 		TogglePictureButton toggleButton;
 
-		projectionWidthSlider = addButton(new NamedSlider(Utils.localize("gui.securitycraft:projector.width", tileEntity.getProjectionWidth()).getColoredString(), blockName, left, topPos + 47, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.width", "").getColoredString(), "", ProjectorBlockEntity.MIN_WIDTH, ProjectorBlockEntity.MAX_WIDTH, tileEntity.getProjectionWidth(), false, true, null, this::sliderReleased));
+		projectionWidthSlider = addButton(new NamedSlider(Utils.localize("gui.securitycraft:projector.width", be.getProjectionWidth()).getColoredString(), blockName, left, topPos + 57, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.width", "").getColoredString(), "", ProjectorBlockEntity.MIN_WIDTH, ProjectorBlockEntity.MAX_WIDTH, be.getProjectionWidth(), false, true, null, this::sliderReleased));
 		projectionWidthSlider.setFGColor(14737632);
 		hoverCheckers[id++] = new StringHoverChecker(projectionWidthSlider, Utils.localize("gui.securitycraft:projector.width.description").getColoredString());
 
-		projectionHeightSlider = addButton(new NamedSlider(Utils.localize("gui.securitycraft:projector.height", tileEntity.getProjectionHeight()).getColoredString(), blockName, left, topPos + 68, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.height", "").getColoredString(), "", ProjectorBlockEntity.MIN_WIDTH, ProjectorBlockEntity.MAX_WIDTH, tileEntity.getProjectionHeight(), false, true, null, this::sliderReleased));
+		projectionHeightSlider = addButton(new NamedSlider(Utils.localize("gui.securitycraft:projector.height", be.getProjectionHeight()).getColoredString(), blockName, left, topPos + 78, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.height", "").getColoredString(), "", ProjectorBlockEntity.MIN_WIDTH, ProjectorBlockEntity.MAX_WIDTH, be.getProjectionHeight(), false, true, null, this::sliderReleased));
 		projectionHeightSlider.setFGColor(14737632);
 		hoverCheckers[id++] = new StringHoverChecker(projectionHeightSlider, Utils.localize("gui.securitycraft:projector.height.description").getColoredString());
 
-		projectionRangeSlider = addButton(new NamedSlider(Utils.localize("gui.securitycraft:projector.range", tileEntity.getProjectionRange()).getColoredString(), blockName, left, topPos + 89, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.range", "").getColoredString(), "", ProjectorBlockEntity.MIN_RANGE, ProjectorBlockEntity.MAX_RANGE, tileEntity.getProjectionRange(), false, true, slider -> {
+		projectionRangeSlider = addButton(new NamedSlider(Utils.localize("gui.securitycraft:projector.range", be.getProjectionRange()).getColoredString(), blockName, left, topPos + 99, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.range", "").getColoredString(), "", ProjectorBlockEntity.MIN_RANGE, ProjectorBlockEntity.MAX_RANGE, be.getProjectionRange(), false, true, slider -> {
 			//show a different number so it makes sense within the world
-			if (tileEntity.isHorizontal())
+			if (be.isHorizontal())
 				slider.setMessage(slider.dispString + Integer.toString((int) Math.round(slider.sliderValue * (slider.maxValue - slider.minValue) + slider.minValue) - 16));
 		}, this::sliderReleased));
-		projectionRangeSlider.setFGColor(14737632);
+		projectionRangeSlider.setFGColor(0xE0E0E0);
 		hoverCheckers[id++] = new StringHoverChecker(projectionRangeSlider, Utils.localize("gui.securitycraft:projector.range.description").getColoredString());
 
-		projectionOffsetSlider = addButton(new NamedSlider(Utils.localize("gui.securitycraft:projector.offset", tileEntity.getProjectionOffset()).getColoredString(), blockName, left, topPos + 110, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.offset", "").getColoredString(), "", ProjectorBlockEntity.MIN_OFFSET, ProjectorBlockEntity.MAX_OFFSET, tileEntity.getProjectionOffset(), false, true, null, this::sliderReleased));
-		projectionOffsetSlider.setFGColor(14737632);
+		projectionOffsetSlider = addButton(new NamedSlider(Utils.localize("gui.securitycraft:projector.offset", be.getProjectionOffset()).getColoredString(), blockName, left, topPos + 120, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.offset", "").getColoredString(), "", ProjectorBlockEntity.MIN_OFFSET, ProjectorBlockEntity.MAX_OFFSET, be.getProjectionOffset(), false, true, null, this::sliderReleased));
+		projectionOffsetSlider.setFGColor(0xE0E0E0);
 		hoverCheckers[id++] = new StringHoverChecker(projectionOffsetSlider, Utils.localize("gui.securitycraft:projector.offset.description").getColoredString());
 
 		//@formatter:off
-		toggleButton = addButton(new TogglePictureButton(left, topPos + 26, 20, 20, TEXTURE, new int[] {176, 192}, new int[] {0, 0}, 2, 2, b -> {
+		toggleButton = addButton(new TogglePictureButton(left + sliderWidth - 20, topPos + 36, 20, 20, TEXTURE, new int[]{176, 192}, new int[]{0, 0}, 2, 2, b -> {
 			//@formatter:on
-			tileEntity.setHorizontal(!tileEntity.isHorizontal());
+			be.setHorizontal(!be.isHorizontal());
 			projectionRangeSlider.updateSlider();
-			SecurityCraft.channel.sendToServer(new SyncProjector(tileEntity.getBlockPos(), tileEntity.isHorizontal() ? 1 : 0, DataType.HORIZONTAL));
+			SecurityCraft.channel.sendToServer(new SyncProjector(be.getBlockPos(), be.isHorizontal() ? 1 : 0, DataType.HORIZONTAL));
 		}));
-		toggleButton.setCurrentIndex(tileEntity.isHorizontal() ? 1 : 0);
+		toggleButton.setCurrentIndex(be.isHorizontal() ? 1 : 0);
 		hoverCheckers[id++] = new StringHoverChecker(toggleButton, Arrays.asList(Utils.localize("gui.securitycraft:projector.vertical").getColoredString(), Utils.localize("gui.securitycraft:projector.horizontal").getColoredString()));
 		projectionRangeSlider.updateSlider();
 
 		slotHoverChecker = new StringHoverChecker(topPos + 22, topPos + 39, leftPos + 78, leftPos + 95, SLOT_TOOLTIP);
+
+		children.add(stateSelector = new StateSelector(menu, title, leftPos - 190, topPos + 7, 0, 197, 0, -2.85F, -0.45F));
+		stateSelector.init(minecraft, width, height);
 	}
 
 	@Override
@@ -94,7 +104,7 @@ public class ProjectorScreen extends ContainerScreen<ProjectorMenu> {
 				renderTooltip(thc.getName(), mouseX, mouseY);
 		}
 
-		if (slotHoverChecker.checkHover(mouseX, mouseY) && menu.te.isEmpty())
+		if (slotHoverChecker.checkHover(mouseX, mouseY) && menu.be.isEmpty())
 			renderTooltip(slotHoverChecker.getName(), mouseX, mouseY);
 	}
 
@@ -109,6 +119,9 @@ public class ProjectorScreen extends ContainerScreen<ProjectorMenu> {
 		RenderSystem.color4f(1.0F, 1.0F, 1.0F, 1.0F);
 		minecraft.getTextureManager().bind(TEXTURE);
 		blit(leftPos, topPos, 0, 0, imageWidth, imageHeight);
+
+		if (stateSelector != null)
+			stateSelector.render(mouseX, mouseY, partialTicks);
 	}
 
 	@Override
@@ -128,27 +141,53 @@ public class ProjectorScreen extends ContainerScreen<ProjectorMenu> {
 		return super.mouseReleased(mouseX, mouseY, button);
 	}
 
+	@Override
+	public boolean mouseDragged(double mouseX, double mouseY, int button, double dragX, double dragY) {
+		if (stateSelector != null && stateSelector.mouseDragged(mouseX, mouseY, button, dragX, dragY))
+			return true;
+
+		return super.mouseDragged(mouseX, mouseY, button, dragX, dragY);
+	}
+
+	@Override
+	public List<Rectangle2d> getExtraAreas() {
+		if (stateSelector != null)
+			return stateSelector.getGuiExtraAreas();
+		else
+			return new ArrayList<>();
+	}
+
+	@Override
+	public void onClose() {
+		super.onClose();
+
+		if (stateSelector.getState() != null) {
+			be.setProjectedState(stateSelector.getState());
+			SecurityCraft.channel.sendToServer(new SyncProjector(be.getBlockPos(), stateSelector.getState()));
+		}
+	}
+
 	public void sliderReleased(NamedSlider slider) {
 		int data = 0;
 		DataType dataType = DataType.INVALID;
 
 		if (slider == projectionWidthSlider) {
-			tileEntity.setProjectionWidth(data = slider.getValueInt());
+			be.setProjectionWidth(data = slider.getValueInt());
 			dataType = DataType.WIDTH;
 		}
 		else if (slider == projectionHeightSlider) {
-			tileEntity.setProjectionHeight(data = slider.getValueInt());
+			be.setProjectionHeight(data = slider.getValueInt());
 			dataType = DataType.HEIGHT;
 		}
 		else if (slider == projectionRangeSlider) {
-			tileEntity.setProjectionRange(data = slider.getValueInt());
+			be.setProjectionRange(data = slider.getValueInt());
 			dataType = DataType.RANGE;
 		}
 		else if (slider == projectionOffsetSlider) {
-			tileEntity.setProjectionOffset(data = slider.getValueInt());
+			be.setProjectionOffset(data = slider.getValueInt());
 			dataType = DataType.OFFSET;
 		}
 
-		SecurityCraft.channel.sendToServer(new SyncProjector(tileEntity.getBlockPos(), data, dataType));
+		SecurityCraft.channel.sendToServer(new SyncProjector(be.getBlockPos(), data, dataType));
 	}
 }