+++ src/main/java/net/geforcemods/securitycraft/util/IHasExtraAreas.java
@@ -1,0 +1,9 @@
+package net.geforcemods.securitycraft.util;
+
+import java.util.List;
+
+import net.minecraft.client.renderer.Rectangle2d;
+
+public interface IHasExtraAreas {
+	public List<Rectangle2d> getExtraAreas();
+}